refactor: 온보딩 프레젠테이션 레이어 패키지 구조 개선 및 정리
- 하위 패키지(`profile`, `welcome`, `terms`, `component`)에 분산되어 있던 주요 화면 및 컴포넌트를 상위 패키지(`presentation`)로 이동하여 구조 단순화
- 대상 파일: `OnboardingProfileScreen`, `WelcomeScreen`, `OnboardingTermsScreen`, `OnboardingStepScaffold`
- 패키지 위치 변경에 따른 프로젝트 내 임포트 경로 및 관련 파일 업데이트
- `SignUpScreen.kt` 내 사용되지 않는 `SIGN_UP_TOTAL_STEPS` 상수 제거 및 불필요한 임포트 정리
