+++ feature/onboarding/presentation/src/main/java/com/afternote/feature/onboarding/presentation/navigation/OnboardingNavGraph.kt
@@ -14,15 +14,15 @@ import androidx.navigation.compose.composable
 import androidx.navigation.navigation
 import com.afternote.core.domain.usecase.auth.LoginType
 import com.afternote.core.ui.Route
+import com.afternote.feature.onboarding.presentation.OnboardingProfileScreen
+import com.afternote.feature.onboarding.presentation.OnboardingTermsScreen
+import com.afternote.feature.onboarding.presentation.WelcomeScreen
 import com.afternote.feature.onboarding.presentation.login.LoginViewModel
 import com.afternote.feature.onboarding.presentation.login.component.LoginScreen
-import com.afternote.feature.onboarding.presentation.profile.OnboardingProfileScreen
 import com.afternote.feature.onboarding.presentation.signup.SignUpPasswordScreen
 import com.afternote.feature.onboarding.presentation.signup.SignUpResidentNumberScreen
 import com.afternote.feature.onboarding.presentation.signup.SignUpScreen
 import com.afternote.feature.onboarding.presentation.signup.SignUpViewModel
-import com.afternote.feature.onboarding.presentation.terms.OnboardingTermsScreen
-import com.afternote.feature.onboarding.presentation.welcome.WelcomeScreen
 
 /**
  * 온보딩 피처의 네비게이션 그래프.
