+++ feature/onboarding/presentation/src/main/java/com/afternote/feature/onboarding/presentation/signup/SignUpScreen.kt
@@ -17,10 +17,8 @@ import com.afternote.core.ui.AfternoteTextField
 import com.afternote.core.ui.TextFieldType
 import com.afternote.core.ui.theme.AfternoteDesign
 import com.afternote.core.ui.theme.AfternoteTheme
+import com.afternote.feature.onboarding.presentation.OnboardingStepScaffold
 import com.afternote.feature.onboarding.presentation.R
-import com.afternote.feature.onboarding.presentation.component.OnboardingStepScaffold
-
-internal const val SIGN_UP_TOTAL_STEPS = 4
 
 @Composable
 fun SignUpScreen(
